added comments to describe the functions and processes

+++ KeyPadBranch/keypad.c
@@ -8,19 +8,22 @@
 Copyright (C) 2022. All rights reserved.
 ******************************************************************************************************************************
 */
+// include librarys
 #include "keypad.h"
 #include "../standard_types.h"
 #include "../macros.h"
-
+// initializing ports and pins needed by keypad
 void Init_Ports(void)
 {
 	// clocking both portC and portD
 	SYSCTL_RCGCGPIO_R |= 0x0C;
+	// wait until Port C & D clocking 
 	while((SYSCTL_PRGPIO_R &0x0C) == 0);
+	// initiallizing port C & D
 	Init_PortC();
 	Init_PortD();
 }
-
+// initiallizing port C
 void Init_PortC(void)
 {
 	GPIO_PORTC_CR_R  		|= 	GPIO_PORTC_PIN4_7;
@@ -31,7 +34,7 @@ void Init_PortC(void)
 	GPIO_PORTC_DEN_R 		|= 	GPIO_PORTC_PIN4_7;
 	GPIO_PORTC_DIR_R 		|= 	GPIO_PORTC_PIN4_7;
 }
-
+// initiallizing port D
 void Init_PortD()
 {
 	GPIO_PORTD_CR_R 		|=  GPIO_PORTD_PIN0_3;
@@ -41,7 +44,7 @@ void Init_PortD()
 	GPIO_PORTD_DEN_R 		|=  GPIO_PORTD_PIN0_3;
 	GPIO_PORTD_DIR_R 		&= ~GPIO_PORTD_PIN0_3;
 }
-
+// return value of corresponding pressed key
 char KeyPad_Get_Input(void)
 {
 	// Values of Buttons in KeyPad
@@ -50,11 +53,11 @@ char KeyPad_Get_Input(void)
 																			{ '4' , '5' , '6' , 'B'}, 	// row_1
 																			{ '7' , '8' , '9' , 'C'}, 	// row_2
 																			{ '*' , '0' , '#' , 'D'} };	// row_3
-
+// waiting pressing key
 	while(1)
 	{ // initialization the counter of row and columns
-		uint32_t row;
-		uint32_t column;
+		uint8_t row;
+		uint8_t column;
 		// looping and set each row high individually
 		for( row = KEYPAD_ROW_START_INDEX ; row < KEYPAD_ROW_LAST_INDEX ; row++) 
 		{
@@ -64,8 +67,10 @@ char KeyPad_Get_Input(void)
 			// with only one certain row high, looping on each column to see which one is pressed
 			for( column = KEYPAD_COLUMN_START_INDEX ; column < KEYPAD_COLUMN_LAST_INDEX ; column++)
 			{
+				// check if a key is pressed and search the column of the pressed key 
 				if((GPIO_PORTC_DATA_R & GPIO_PORTC_PIN4_7) == KEYPAD_FIRST_COLUMN << column)
 				{
+					// return the corresponding value of pressed key
 					return Values[row][column];
 				}
 			}
